+++ CHANGES.rst
@@ -27,6 +27,7 @@ Improvements
   room bookings (:issue:`5730`, :pr:`5740`)
 - Add day of the week to room booking details modal and timeline (:issue:`5718`,
   :pr:`5743`)
+- Allow acceptance and rejection of editables in the editable list (:pr:`5721`)
 
 Bugfixes
 ^^^^^^^^